Move /ui/main/src/app/business/store one level up (#7788)
