+++ ui/main/src/app/business/view/timeline/timeline.view.ts
@@ -9,9 +9,9 @@
 
 import {Circle, Circles} from './circles';
 import {XAxis} from './xaxis';
-import {FilteredLightCardsStore} from 'app/business/store/lightcards/lightcards-feed-filter-store';
+import {FilteredLightCardsStore} from '../../../store/lightcards/lightcards-feed-filter-store';
 import {Observable, Subject, takeUntil} from 'rxjs';
-import {OpfabStore} from 'app/business/store/opfabStore';
+import {OpfabStore} from '../../../store/opfabStore';
 import {DateTimeFormatterService} from 'app/services/dateTimeFormatter/DateTimeFormatterService';
 
 export class TimelineView {
