+++ ui/main/src/app/components/core/application-loading/application-loading.component.ts
@@ -13,7 +13,7 @@ import {firstValueFrom, Subject, tap} from 'rxjs';
 import {ActivityAreaChoiceAfterLoginComponent} from './activityarea-choice-after-login/activityarea-choice-after-login.component';
 import {AppLoadedInAnotherTabComponent} from './app-loaded-in-another-tab/app-loaded-in-another-tab.component';
 import {OpfabEventStreamServer} from '@ofServices/events/server/OpfabEventStreamServer';
-import {CurrentUserStore} from 'app/business/store/current-user.store';
+import {CurrentUserStore} from '../../../store/current-user.store';
 import {AuthService} from 'app/authentication/auth.service';
 import {AuthenticationMode} from 'app/authentication/auth.model';
 import {RemoteLoggerServer} from '@ofServices/logs/server/RemoteLoggerServer';
