+++ ui/main/src/app/components/feed/components/card-list/filters/feed-filter/feed-filter.component.ts
@@ -15,9 +15,9 @@ import * as _ from 'lodash-es';
 import {FilterType} from '@ofModel/feed-filter.model';
 import {UserPreferencesService} from '@ofServices/userPreferences/UserPreferencesService';
 import {MessageLevel} from '@ofServices/alerteMessage/model/Message';
-import {FilteredLightCardsStore} from 'app/business/store/lightcards/lightcards-feed-filter-store';
+import {FilteredLightCardsStore} from '../../../../../../store/lightcards/lightcards-feed-filter-store';
 import {AlertMessageService} from '@ofServices/alerteMessage/AlertMessageService';
-import {OpfabStore} from 'app/business/store/opfabStore';
+import {OpfabStore} from '../../../../../../store/opfabStore';
 import {MultiSelect, MultiSelectOption} from 'app/components/share/multi-select/model/MultiSelect';
 import {ProcessesService} from '@ofServices/processes/ProcessesService';
 import {UsersService} from '@ofServices/users/UsersService';
