+++ ui/main/src/app/components/feed/components/card-list/filters/filters.component.ts
@@ -9,7 +9,7 @@
 
 import {Component, Input, OnInit, Output} from '@angular/core';
 import {ConfigService} from 'app/services/config/ConfigService';
-import {OpfabStore} from 'app/business/store/opfabStore';
+import {OpfabStore} from '../../../../../store/opfabStore';
 import {Subject} from 'rxjs';
 import {FeedFilterAndSortIconsComponent} from './feed-filter-and-sort-icons/feed-filter-and-sort-icons.component';
 import {FeedSearchComponent} from './feed-search/feed-search.component';
