+++ ui/main/src/app/components/feed/feed.component.ts
@@ -11,11 +11,11 @@ import {Component, OnDestroy, OnInit} from '@angular/core';
 import {Observable, Subject} from 'rxjs';
 import {LightCard} from '@ofModel/light-card.model';
 import {delay, map, takeUntil} from 'rxjs/operators';
-import {FilteredLightCardsStore} from 'app/business/store/lightcards/lightcards-feed-filter-store';
+import {FilteredLightCardsStore} from '../../store/lightcards/lightcards-feed-filter-store';
 import {ConfigService} from 'app/services/config/ConfigService';
 import {ActivatedRoute, RouterOutlet} from '@angular/router';
 import {SelectedCardService} from '@ofServices/selectedCard/SelectedCardService';
-import {OpfabStore} from 'app/business/store/opfabStore';
+import {OpfabStore} from '../../store/opfabStore';
 import {TimeLineComponent} from './components/time-line/time-line.component';
 import {NgIf, AsyncPipe} from '@angular/common';
 import {PinnedCardsComponent} from './components/pinned-cards/pinned-cards.component';
