+++ ui/main/src/app/components/monitoring/components/cell-renderers/responses-cell-renderer.component.ts
@@ -13,7 +13,7 @@ import {EntitiesService} from '@ofServices/entities/EntitiesService';
 import {ICellRendererAngularComp} from 'ag-grid-angular';
 import {ICellRendererParams} from 'ag-grid-community';
 import {filter, Subject, takeUntil} from 'rxjs';
-import {OpfabStore} from 'app/business/store/opfabStore';
+import {OpfabStore} from '../../../../store/opfabStore';
 import {NgIf, NgFor, NgStyle} from '@angular/common';
 import {NgbPopover} from '@ng-bootstrap/ng-bootstrap';
