+++ ui/main/src/app/components/monitoring/components/monitoring-filters/monitoring-filter-builder.ts
@@ -11,8 +11,8 @@ import {Injectable} from '@angular/core';
 import {Filter} from '@ofModel/feed-filter.model';
 import {LightCard} from '@ofModel/light-card.model';
 import {ProcessesService} from '@ofServices/processes/ProcessesService';
-import {FilteredLightCardsStore} from 'app/business/store/lightcards/lightcards-feed-filter-store';
-import {OpfabStore} from 'app/business/store/opfabStore';
+import {FilteredLightCardsStore} from '../../../../store/lightcards/lightcards-feed-filter-store';
+import {OpfabStore} from '../../../../store/opfabStore';
 
 @Injectable({
     providedIn: 'root'
