+++ ui/main/src/app/server/interceptors.service.ts
@@ -10,7 +10,7 @@
 import {Injectable} from '@angular/core';
 import {HttpEvent, HttpHandler, HttpInterceptor, HttpRequest} from '@angular/common/http';
 import {Observable} from 'rxjs';
-import {CurrentUserStore} from 'app/business/store/current-user.store';
+import {CurrentUserStore} from '../store/current-user.store';
 
 @Injectable({
     providedIn: 'root'
