+++ ui/main/src/app/services/notifications/SystemNotificationService.ts
@@ -15,7 +15,7 @@ import {LogOption, LoggerService as logger} from 'app/services/logs/LoggerServic
 import {filter} from 'rxjs/operators';
 import {MessageLevel} from '@ofServices/alerteMessage/model/Message';
 import {AlertMessageService} from '../alerteMessage/AlertMessageService';
-import {OpfabStore} from 'app/business/store/opfabStore';
+import {OpfabStore} from '../../store/opfabStore';
 import {NavigationService} from '../navigation/NavigationService';
 import {NotificationDecision} from './NotificationDecision';
