+++ ui/main/src/app/services/userSettings/UserSettingsService.ts
@@ -10,7 +10,7 @@
 import {Observable} from 'rxjs';
 import {UserSettingsServer} from './server/UserSettingsServer';
 import {LogOption, LoggerService as logger} from 'app/services/logs/LoggerService';
-import {CurrentUserStore} from '../../business/store/current-user.store';
+import {CurrentUserStore} from '../../store/current-user.store';
 import {ServerResponse} from 'app/business/server/serverResponse';
 
 export class UserSettingsService {
