+++ ui/main/src/app/store/lightcards/lightcards-text-filter.ts
